# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/pos_payments:
    get:
      tags:
        - PosPayments
      summary: List PosPayments
      description: Get a paginated list of PosPayments with optional filtering and sorting
      operationId: listPosPayments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosPaymentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PosPayments
      summary: Create PosPayment
      description: Create a new PosPayment
      operationId: createPosPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePosPaymentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosPayment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/pos_opening_entries:
    get:
      tags:
        - PosOpeningEntries
      summary: List PosOpeningEntries
      description: Get a paginated list of PosOpeningEntries with optional filtering and sorting
      operationId: listPosOpeningEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosOpeningEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PosOpeningEntries
      summary: Create PosOpeningEntry
      description: Create a new PosOpeningEntry
      operationId: createPosOpeningEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePosOpeningEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosOpeningEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PosPayment:
      type: object
      required:
        - id
        - company_id
        - pos_invoice_id
        - payment_method
        - amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        pos_invoice_id:
          type: string
          format: uuid
        payment_method:
          type: string
        amount:
          type: string
        reference_no:
          nullable: true
          type: string
          maxLength: 140
        payment_entry_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreatePosPaymentRequest:
      type: object
      required:
        - company_id
        - pos_invoice_id
        - payment_method
        - amount
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        pos_invoice_id:
          type: string
          format: uuid
        payment_method:
          type: string
        amount:
          type: string
        reference_no:
          nullable: true
          type: string
          maxLength: 140
        payment_entry_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    PosPaymentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PosPayment'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PosOpeningEntry:
      type: object
      required:
        - id
        - company_id
        - pos_profile_id
        - cashier_party_id
        - opened_at
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        pos_profile_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        cashier_party_id:
          type: string
          format: uuid
        opened_at:
          type: string
          format: date-time
        opening_balances:
          nullable: true
          type: object
        status:
          type: string
        metadata:
          type: object
...
    CreatePosOpeningEntryRequest:
      type: object
      required:
        - company_id
        - pos_profile_id
        - cashier_party_id
        - opened_at
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        pos_profile_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        cashier_party_id:
          type: string
          format: uuid
        opened_at:
          type: string
          format: date-time
        opening_balances:
          nullable: true
          type: object
        status:
          type: string
        metadata:
          type: object
...
    PosOpeningEntryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PosOpeningEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []